Fixes linking issues with VS2015 + Boost 1.64.
Reimplemented PyStrand as an shared_ptr<qi::Strand> which
simplifies the passing of that object in boost::python::object
and it's extraction. This change avoids the linking issue
which only happen when we try to make boost::python do the ownership
automatically. The code is now clearer on the ownership semantics.

Change-Id: I1f6a1dcab4ee8fb90cec3c1e54bcf48886f1a51f
Reviewed-on: http://gerrit.aldebaran.lan/88450
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: gerrit