PyQiObject: Adds support for object identification.
It is now possible to write in python `a == b`, where `a` and `b` are objects.
This syntax tests that the two objects share the same underlying pointer.

It is thus more an identity test than an equality test. For example with two
distinct objects constructed with the same parameters, `a == b` is false.

All relational operators are provided: `!=`, `<`, `<=`, `>`, `>=`

Change-Id: Iad24c68865ed704eae9ff8797955eb4705adf16a
Reviewed-on: http://gerrit.aldebaran.lan/89060
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: gerrit