Stores and maintains ObjectUid of Python objects going through qi functions. #44830
Fixes #44830.

Each time a python object was passed (as a `qi::Object`) through libqi,
a new `qi::AnyObject` was created with a new uid (as per default).
This means that each call with the same object parameter sent
different uids, making impossible to identify objects through this means.

This change stores in the python object the ObjectUid generated
the first time that object is passed through libqi. Any other times
that ObjectUid value will be reused instead of creating a new one.

Requires these fixes to be in libqi for this change to work:
 - http://gerrit.aldebaran.lan/96488
 - http://gerrit.aldebaran.lan/96489
 - http://gerrit.aldebaran.lan/96490

Change-Id: Iba4382e98013321853a42e666a33ffed64a58ff4
Signed-off-by: Joël LAMOTTE <jlamotte@softbankrobotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/96491
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>