Ports to python3 and pybind11 and redoes the build system.
This patch ports the base code to Python3 (3.5+) only, ports the C++
code from boost::python to pybind11 and redoes the build system of the
library.

Boost python is getting outdated and is not maintained as much as we
would need, so we decided to use pybind11. This has the impact
of changing a lot of C++ code in the library, although the code should
behave the same, aside from changes that came with Python3.

This patch also cleans up the Python code so that pep8 and pyflakes do
not report any error or warning.

Finally, it redoes the entire build system of the library. The
CMakeLists now requires CMake 3.17+ and uses a more modern approach to
find dependencies and define our artifacts. The setup.py now uses
scikit-build to interface with CMake and automatically build a wheel
package.

Also fixes #45921.

Change-Id: Id491a9d4a14b6bbf5ec60bf767bf5e93d69ad993
Reviewed-on: http://gerrit2.aldebaran.lan/424
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>