types: Handles conversion of `LogLevel` python values into C++. #46315
When the library was ported to Python 3 and rewritten using pybind11, a
new enumeration type was introduced in Python, the `LogLevel`.

Unfortunately, the function handling conversion of Python objects into
typed C++ values did not handle `LogLevel` values correctly.
Furthermore, in libqi, the `LogLevel` type is not registered in the qi
type system, but when registered, enum types are always treated as
`int` values. This forces us to handle it as an `int`.

Change-Id: I270ec7ea9ac407c2b6b5baab31d0bc78903cfcc4
Reviewed-on: http://gerrit2.aldebaran.lan/1233
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>