cmake: Replaces CACHE variables with overridable variables.
This patch replaces every CACHE variable that has a hard-coded default
value with a overridable variable (using a new function that automatizes
its creation).

An overridable variable is a variable that uses a default value if it
doesn't already exists, either in the cache or in the environment.

Doing this enables us to update the hard-coded values by changing the
CMake files, consequently updating any existing configuration that
doesn't have an overriden value for these variables (therefore using
the default values).

Change-Id: I740ee98c0d680be4d86940968af320bc1e1a634e
Reviewed-on: http://gerrit2.aldebaran.lan/1384
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>