misc: Fixes Python version check code.
The previous code was checking if one of the tuple was less than the
other, but the tuples contain 3 strings (e.g. `('3', '5', '0')`).
Unfortunately, string comparison does not behave like we naively
thought it did, as for instance, '5' is not less than '10'. This was
causing a failure for the check for Python 3.10.

This patch rewrites this check using the `packaging.version` module
which implements correct version comparison according to PEP 440.

Change-Id: Icacb8fb0a2c0cbaa0b159a9269a3fbde152ee412
Reviewed-on: http://gerrit2.aldebaran.lan/1613
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>