cmake.install: Uses `patchelf` when available to set the runtime path of external shared libraries.
This patch uses the `patchelf` tool to set the runtime path of external
shared libraries wrapped in the package to '$ORIGIN' or '@loader_path'.
These special values refer to the path of the shared library itself,
depending on the system.

All the shared libraries that are wrapped in the package are extracted
in the same directory when the package is installed.

This allows us to ensure that any shared library that gets
wrapped in the package can find its dependencies also wrapped in
the package.

Change-Id: I5e1ec841942e20a7dcefa5d90a67073635709272
Reviewed-on: http://gerrit2.aldebaran.lan/1622
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>