build: Upgrades version of pybind11 from v2.5.0 to v2.9.0.
The public constructor of `pybind11::module` is now deprecated. This
patch replaces its only use by a call to `PYBIND11_EMBEDDED_MODULE`
followed by a `pybind11::import(...)`.

The constructor of a `pybind11::object` subclass from a
`pybind11::object` from the wrong type now throws an exception. See
https://github.com/pybind/pybind11/pull/2349.

We had one case of this happening where we tried constructing a `bytes`
from a `None`. We now first construct an `object`, then test if it's
`None` before trying to convert it to `bytes`.

Change-Id: I4aa2a1b4a49d3024f203bdd5c6f9c4ee86bfc9c4
Reviewed-on: http://gerrit2.aldebaran.lan/1760
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>