build: Upgrades version of googletest to live at head.
The tagged versions of Googletest are considered obsolete. The
recommended version is the one targeted by the `main` branch.

This patch adds an include to `optional_io.hpp`, because a change
in googletest caused one of our tests to try to output a
`boost::optional`.


Integration breaking change: the GOOGLETEST_VERSION CMake variable is no
longer honored.

Change-Id: Ie816821af53a79842ddad4f2e5c56e26a2e20782
Reviewed-on: http://gerrit2.aldebaran.lan/1761
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>