qi.pyguard: Fixes segfault at Python interpreter finalization due to GIL bad manipulation. #47167
A segfault occurs during the finalization of the Python interpreter when an
object has a destructor that uses `GILRelease`. The constructor of `GILRelease`
effectively releases the GIL but the destructor disables its acquisition
because the interpreter is finalizing. When the call returns to the
interpreter, the GIL stays released and this leads to crashes or errors when
the interpreter tries to make Python calls.

This patch rewrites the `GILRelease` type so that during finalization, the
object does nothing (therefore the GIL is not released).