build: Rewrites project build definition. #SW-159
This patch rewrites the definition of the build of libqi-python in order to remove the dependency to `qibuild` and `qi toolchains` and instead use more standard tools and conventions.

It completely refactors the `CMakeLists.txt` files to use an approach closer to what is referred to as "modern CMake". The project now requires CMake 3.23 or greater.

The project defines a single install components, the `Module` component.
Tests are now declared to CMake using `add_test` and the CTest module is used. This means that tests can now be executed with the `ctest` command line tool.

This patch also adds the support of Conan, to download the project dependencies and to create a Conan package of the project. Please refer to the README for basic usage.

It also rewrites the Python module project definition using PEP517 and `scikit-build-core` as a build backend.