Add Gitlab CI to build wheels.
This patch adds Gitlab CI to the project to build wheels for Linux for each supported version of Python.

This build can be manually triggered for any commit, but is automatic when a "qi-python-vXXX" tag is pushed.
Furthermore, when a tag is pushed on GitLab, wheels are published in the package registry and a release is created.